Revert "Merge remote-tracking branch 'origin/orientation2' into orientation2"
This reverts commit 2b6e899cb0631273987e663bd691bd0962ea737b.

+++ src/utils_dijkstra.c
@@ -19,5 +19,5 @@ int get_distance_x_y(borne *bDep, borne *bArr) {
 
 // returns the charging time of a vehicule at a given borne
 int get_charging_time(borne* b,vehicule* v, int max_time_charging){
-    return MIN((int)((float)v->capacity / (float)b->power)*60.0, max_time_charging);
+    return MIN((int)(v->capacity / b->power)*60, max_time_charging);
 }
